find median from data stream

+++ leetcode/295/MedianFinder.java
@@ -36,7 +36,7 @@ class MedianFinder {
         if (maxHeap.size() == minHeap.size()) {
             return (double) (minHeap.peek() + maxHeap.peek()) / 2;
         } else {
-            return maxHeap.peek();
+            return (double)maxHeap.peek();
         }
     }
 };
